+++ docs/ChangeLog.md
@@ -484,6 +484,7 @@
 - util.hook: workaround bash-5.2 bug of nested read by `WINCH` `#D1981` a5b10e8
 - edit: always adjust the terminal states with `bind -x` (reported by linwaytin) `#D1983` 5d14cf1
 - syntax: suppress brace expansions in designated array initialization in Bash 5.3 `#D1989` 1e7b884
+- progcomp: work around slow `nix` completion `#D1997` xxxxxxx
 
 ## Test